+++ cmds/command_lists.go
@@ -15,13 +15,16 @@ var (
 	_listInternalID = ListsCmd.Flag.String("i", cfg.DefaultConfig().InternalID, consts.ListIDUsage)
 	_listSort       = ListsCmd.Flag.String("s", cfg.DefaultConfig().CommentsSort, consts.ListCommentSortUsage)
 	_listLikeRemove = ListsCmd.Flag.Bool("remove", cfg.DefaultConfig().Remove, consts.ListLikeRemoveUsage)
+	_listType       = ListsCmd.Flag.String("t", cfg.DefaultConfig().ListsType, consts.ListTypeUsage)
+	_listReason     = ListsCmd.Flag.String("r", cfg.DefaultConfig().Reason, consts.ReasonUsage)
+	_listMessage    = ListsCmd.Flag.String("message", cfg.DefaultConfig().Msg, consts.ReportMsgUsage)
 )
 
 // ListsCmd returns movies and episodes that a user has watched, sorted by most recent.
 var ListsCmd = &Command{
 	Name:    "lists",
 	Usage:   "",
-	Summary: "Returns data about lists: trending, popular, list, likes, like, items, comments.",
+	Summary: "Returns data about lists: trending, popular, list, likes, like, items, comments, report.",
 	Help:    `lists command`,
 }
 
@@ -39,11 +42,12 @@ func listsFunc(cmd *Command, _ ...string) error {
 		"like":     handlers.ListsLikeHandler{},
 		"items":    handlers.ListsItemsHandler{},
 		"comments": handlers.ListsCommentsHandler{},
+		"report":   handlers.ListsReportHandler{},
 	}
 
 	handler, err := cmd.common.GetHandlerForMap(options.Action, allHandlers)
 
-	validActions = []string{"trending", "popular", "list", "likes", "like", "items", "comments"}
+	validActions = []string{"trending", "popular", "list", "likes", "like", "items", "comments", "report"}
 	if err != nil {
 		cmd.common.GenActionsUsage(cmd.Name, validActions)
 		return nil