+++ consts/glob.go
@@ -79,6 +79,8 @@ const (
 	LanguageUsage                = "allow to overwrite language"
 	ListCommentSortUsage         = "allow to overwrite comments sort"
 	ListIDUsage                  = "allow to export a specific custom list"
+	ListTypeUsage                = "list type for trending/popular, or item types for items ie: movie,show"
+	ListReportPending            = "list %s already has a pending report from this user"
 	ListItem                     = "  - %s\n"
 	ItemIDUsage                  = "allow to overwrite item_id"
 	ListItemIDUsage              = "allow to overwrite list_item_id"