+++ handlers/lists_report_handler.go
@@ -1,0 +1,40 @@
+// Package handlers used to handle module actions
+package handlers
+
+import (
+	"errors"
+	"fmt"
+
+	"github.com/mfederowicz/trakt-sync/consts"
+	"github.com/mfederowicz/trakt-sync/internal"
+	"github.com/mfederowicz/trakt-sync/printer"
+	"github.com/mfederowicz/trakt-sync/str"
+)
+
+// ListsReportHandler struct for handler
+type ListsReportHandler struct{ common CommonLogic }
+
+// Handle to handle lists: report action
+func (h ListsReportHandler) Handle(options *str.Options, client *internal.Client) error {
+	if len(options.InternalID) == consts.ZeroValue {
+		return errors.New(consts.EmptyListIDMsg)
+	}
+	if len(options.Reason) == consts.ZeroValue {
+		return errors.New(consts.EmptyReasonMsg)
+	}
+	if err := h.common.ValidReason(options); err != nil {
+		return err
+	}
+
+	report := &str.ListReport{Reason: &options.Reason}
+	if len(options.Msg) > consts.ZeroValue {
+		report.Message = &options.Msg
+	}
+
+	if _, err := client.Lists.ReportList(client.BuildCtxFromOptions(options), &options.InternalID, report); err != nil {
+		return fmt.Errorf("report error: %w", err)
+	}
+
+	printer.Printf("reported list %s\n", options.InternalID)
+	return nil
+}